Revert "Plan 9 from Bell Labs Fourth Edition 2006-05-22"
This reverts commit ec59a3ddbfceee0efe34584c2c9981a5e5ff1ec4.

+++ sys/src/cmd/unix/drawterm/gui-x11/xmem.h
@@ -31,7 +31,7 @@ struct Xmem
 	int	dirty;
 	Rectangle dirtyr;
 	Rectangle r;
-	uintptr pc;	/* who wrote into xi */
+	ulong pc;	/* who wrote into xi */
 };
 
 extern	int		xtblbit;
